First tests for the data service
This closes #5, since it checks that the exception is thrown.

It's related to #4, since it contains scaffolding for the YAML class.

It's related to #3, since it starts the abstract classes and a concrete one for storing data

It's related eventually to #1, since this will be used to store check-in data, as reflected in the "add" signature